Rounding errors

+++ include/GUI.hpp
@@ -16,7 +16,7 @@ public:
   GUIContainer *get_root_container() const;
   void add_container(GUIContainer *container);
   void add_text(GUIText *text);
-  void on_window_resize(int width, int height);
+  void on_window_resize(float width, float height);
   void on_mouse_move(float x1, float y1, float x2, float y2);
   void on_mouse_down(float x, float y);
   void on_mouse_up(float x, float y);
